+++ packages/dnd/src/components/index.ts
@@ -6,11 +6,11 @@ export * from './DragHandle';
 export * from './Draggable';
 export * from './DraggableBlock';
 export * from './DraggableBlockToolbar';
+export * from './DraggableBlockToolbarWrapper';
 export * from './DraggableDropline';
 export * from './DraggableGutterLeftProps';
 export * from './DraggableRoot';
 export * from './useDraggableState';
-export * from './useWithDraggable';
+export * from './useWithDraggableState';
 export * from './withDraggable';
 export * from './Scroller/index';
-export { DraggableBlockToolbarWrapper } from './DraggableBlockToolbarWrapper';
