+++ packages/ui/dnd/src/withPlateDraggable.tsx
@@ -1,19 +1,13 @@
 import { createNodesWithHOC, RenderFunction } from '@udecode/plate-core';
 import { withDraggable, WithDraggableOptions } from '@udecode/plate-dnd';
-import { PlateDraggableProps } from './Draggable.types';
-import { PlateDraggable } from './PlateDraggable';
+import { PlateDraggable, PlateDraggableProps } from './PlateDraggable';
 
 export const withPlateDraggable = (
   Component: RenderFunction<any>,
   options?: WithDraggableOptions<
     Partial<Omit<PlateDraggableProps, 'editor' | 'element' | 'children'>>
   >
-) => {
-  return withDraggable<PlateDraggableProps>(
-    PlateDraggable,
-    Component,
-    options as any
-  );
-};
+) =>
+  withDraggable<PlateDraggableProps>(PlateDraggable, Component, options as any);
 
 export const withPlateDraggables = createNodesWithHOC(withPlateDraggable);
